image-processing-queue/worker/requirements.txt
minio
pika
Pillow
